+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 * [UIINIMP-67](https://folio-org.atlassian.net/browse/UIINIMP-67) Show Number of files queued/File currently being processed in edit form (read-only).
+* [UIINIMP-66](https://folio-org.atlassian.net/browse/UIINIMP-66) Add new fields to view/edit Channel: Harvest URL, and Last harvested.
 
 ## [4.1.2](https://github.com/folio-org/ui-inventory-import/tree/v4.1.2) (2026-06-01)